refactor: msw mentoring 핸들러를 home 핸들러로 경로 변경 #894
